+++ src/utils.js
@@ -17,7 +17,7 @@ function formatTimezone(offset) {
   const offsetValue = Math.abs(offset);
   const hours = Math.floor(offsetValue / 60);
 
-  return (Math.sign(offset) ? '+' : '-')
+  return (Math.sign(offset) >= 0 ? '+' : '-')
     + [hours, offsetValue - hours * 60].map((value) => String(value).padStart(2, 0)).join(':');
 }
 
@@ -49,7 +49,7 @@ function parseDate(value) {
   if (value === null || value === undefined || typeof value === 'boolean') return;
 
   return typeof value === 'string'
-    ? DateTime.fromISO(value)
+    ? DateTime.fromISO(value, {setZone:true})
     : value instanceof Date
       ? DateTime.fromJSDate(value)
       : typeof value === 'number' ? DateTime.fromMillis(value) : value;